feat(base): editing existing blogposts implemented
We are now able to edit existing blogposts through the admin dropdown
menu located on each blogpost. The blogpost id is sent to the page and
prefills the content for editing.

Squash of the below commits:

fix(base): css formatting and flash on top

feat(blog): editing existing blogposts implemented