feat(blog): blogpost drafting
Blogposts can now be stored as drafts in the editor and published when
they are ready. This is implemented with a checkbox to toggle a database
column between True or False.
The /blog now filters to only published posts and all drafts can be seen
in the editor in a card view.

To ease implementation the flask-migrate tool was added to migrate
existing databases for the future. This allows us to modify the columns
of the SQL tables without having to reset the data.

Squash of the following commits:

feat(blog): drafting system implemented

fix(database): app now able to migrate databases with flask-migrate